Make the code more OOP
