--- src/main/java/chatData.txt
@@ -1,2 +1,0 @@
-T0123
-D1newone%2022-02-01
